requirements.txt
# 共享单车数据分析与强化学习调度 - Python依赖
# ====================================================
# 更新日期: 2025-10-29 (Day 4)
# Python版本: 3.12+

# ============================================================================
# 核心数值计算与数据处理
# ============================================================================
numpy>=1.21.0          # 数值计算基础库
pandas>=1.3.0          # 数据处理和分析
scipy>=1.7.0           # 科学计算（Poisson分布等）

# ============================================================================
# 配置管理
# ============================================================================
pyyaml>=5.4.0          # YAML配置文件解析

# ============================================================================
# 网络与图算法
# ============================================================================
networkx>=2.6.0        # 网络流优化（Min-Cost Flow策略）

# ============================================================================
# 可视化
# ============================================================================
matplotlib>=3.4.0      # 基础绘图
seaborn>=0.11.0        # 统计可视化（可选）
# pyecharts>=1.9.0     # 交互式图表（Day 10-12需要）

# ============================================================================
# 强化学习
# ============================================================================
gymnasium>=0.28.0      # OpenAI Gym环境接口（替代gym）
# stable-baselines3>=2.0.0  # RL算法库（Day 7-9需要）
# torch>=2.0.0         # PyTorch（stable-baselines3依赖）

# ============================================================================
# 大数据处理
# ============================================================================
pyspark>=3.2.0         # Spark分布式计算

# ============================================================================
# 机器学习
# ============================================================================
scikit-learn>=1.0.0    # 机器学习工具
xgboost>=1.5.0         # XGBoost需求预测

# ============================================================================
# 表格格式化 (Day 4新增)
# ============================================================================
tabulate>=0.8.0        # Markdown表格生成（评估报告必需）

# ============================================================================
# Web应用 (Day 10-12需要)
# ============================================================================
# flask>=2.0.0         # Web框架
# flask-cors>=3.0.0    # 跨域支持

# ============================================================================
# 开发工具（可选）
# ============================================================================
# pytest>=7.0.0        # 单元测试
# black>=22.0.0        # 代码格式化
# flake8>=4.0.0        # 代码检查
# mypy>=0.990          # 类型检查

# ============================================================================
# 注意事项
# ============================================================================
# 1. 在WSL环境中安装时，可能需要使用 --break-system-packages 标志
#    示例: pip install -r requirements.txt --break-system-packages
#
# 2. 如果遇到权限问题，使用 --user 标志
#    示例: pip install -r requirements.txt --user
#
# 3. 建议使用虚拟环境
#    python3 -m venv venv
#    source venv/bin/activate
#    pip install -r requirements.txt
#
# 4. PySpark可能需要额外配置Java环境
#    确保已安装Java 8或11
#
# 5. stable-baselines3需要PyTorch，可能比较大（几百MB）
#    在Day 7-9训练RL时再安装